chore: Breakup `vane.util.Util`
