+++ vane-core/src/main/java/org/oddlama/vane/core/config/ConfigExtendedMaterialField.java
@@ -1,6 +1,6 @@
 package org.oddlama.vane.core.config;
 
-import static org.oddlama.vane.util.Util.namespaced_key;
+import static org.oddlama.vane.util.StorageUtil.namespaced_key;
 
 import java.lang.reflect.Field;
 import java.util.function.Function;
