+++ vane-portals/src/main/java/org/oddlama/vane/portals/menu/ConsoleMenu.java
@@ -25,7 +25,7 @@ import org.oddlama.vane.portals.event.PortalDestroyEvent;
 import org.oddlama.vane.portals.event.PortalSelectTargetEvent;
 import org.oddlama.vane.portals.event.PortalUnlinkConsoleEvent;
 import org.oddlama.vane.portals.portal.Portal;
-import org.oddlama.vane.util.Util;
+import org.oddlama.vane.util.StorageUtil;
 
 public class ConsoleMenu extends ModuleComponent<Portals> {
 
@@ -100,7 +100,7 @@ public class ConsoleMenu extends ModuleComponent<Portals> {
 			new TranslatedItemStack<>(
 				ctx,
 				"unlink_console",
-				Util.namespaced_key("vane", "decoration_tnt_1"),
+				StorageUtil.namespaced_key("vane", "decoration_tnt_1"),
 				1,
 				"Used to unlink the current console."
 			);
@@ -108,7 +108,7 @@ public class ConsoleMenu extends ModuleComponent<Portals> {
 			new TranslatedItemStack<>(
 				ctx,
 				"unlink_console_confirm_accept",
-				Util.namespaced_key("vane", "decoration_tnt_1"),
+				StorageUtil.namespaced_key("vane", "decoration_tnt_1"),
 				1,
 				"Used to confirm unlinking the current console."
 			);
