+++ vane-portals/src/main/java/org/oddlama/vane/portals/menu/SettingsMenu.java
@@ -19,7 +19,7 @@ import org.oddlama.vane.core.module.ModuleComponent;
 import org.oddlama.vane.portals.Portals;
 import org.oddlama.vane.portals.event.PortalChangeSettingsEvent;
 import org.oddlama.vane.portals.portal.Portal;
-import org.oddlama.vane.util.Util;
+import org.oddlama.vane.util.StorageUtil;
 
 public class SettingsMenu extends ModuleComponent<Portals> {
 
@@ -50,7 +50,7 @@ public class SettingsMenu extends ModuleComponent<Portals> {
 			new TranslatedItemStack<>(
 				ctx,
 				"select_icon",
-				Util.namespaced_key("vane", "decoration_end_portal_orb"),
+				StorageUtil.namespaced_key("vane", "decoration_end_portal_orb"),
 				1,
 				"Used to select the portal's icon."
 			);
