+++ vane-regions/src/main/java/org/oddlama/vane/regions/menu/RegionMenu.java
@@ -22,7 +22,7 @@ import org.oddlama.vane.regions.Regions;
 import org.oddlama.vane.regions.region.Region;
 import org.oddlama.vane.regions.region.RegionGroup;
 import org.oddlama.vane.regions.region.RegionSelection;
-import org.oddlama.vane.util.Util;
+import org.oddlama.vane.util.StorageUtil;
 
 public class RegionMenu extends ModuleComponent<Regions> {
 
@@ -53,7 +53,7 @@ public class RegionMenu extends ModuleComponent<Regions> {
 			new TranslatedItemStack<>(
 				ctx,
 				"delete",
-				Util.namespaced_key("vane", "decoration_tnt_1"),
+				StorageUtil.namespaced_key("vane", "decoration_tnt_1"),
 				1,
 				"Used to delete this region."
 			);
@@ -61,7 +61,7 @@ public class RegionMenu extends ModuleComponent<Regions> {
 			new TranslatedItemStack<>(
 				ctx,
 				"delete_confirm_accept",
-				Util.namespaced_key("vane", "decoration_tnt_1"),
+				StorageUtil.namespaced_key("vane", "decoration_tnt_1"),
 				1,
 				"Used to confirm deleting the region."
 			);
