+++ vane-trifles/src/main/java/org/oddlama/vane/trifles/items/XpBottles.java
@@ -2,7 +2,7 @@ package org.oddlama.vane.trifles.items;
 
 import static org.oddlama.vane.util.PlayerUtil.give_item;
 import static org.oddlama.vane.util.PlayerUtil.remove_one_item_from_hand;
-import static org.oddlama.vane.util.Util.exp_for_level;
+import static org.oddlama.vane.util.Conversions.exp_for_level;
 
 import java.util.ArrayList;
 import java.util.EnumSet;
